ECIES flags cleanup
ECIES has several flags which control behavior.

Two of the flags are only relevant for curves with cofactors, which are
already rare and deprecated, and which will be removed in Botan4.

The CheckMode flag is already ignored completely; when decrypting we always
check that the input point is in the prime order subgroup.

The last flag (SingleHashMode) is actually relevant. This can just be a boolean.

Add new constructors that just have the SingleHashMode toggle, and deprecate the
others with intent to remove them with the rest of the cofactor support.

ISO 18033 has a requirement that (when CheckMode = 0) we verify that the prime
order subgroup and the cofactor are relatively prime. A curve that did not
satisfy this condition could in theory be created, though it would be absurd.
Such a curve can only exist due to an application explicitly creating it; when
decoding explicit curve params we require the cofactor be no larger than 16.
Since we implicitly ignore the setting of CheckMode, there is anyway no reason
to perform the test.